requirements_Version2.txt
openai>=1.30.0
tiktoken
pyyaml
GitPython
pyyaml